测试数据

+++ cfg/v2/testdata/main.go
@@ -5,6 +5,7 @@ import (
 	"log"
 )
 
+//   -a=101 -aa=101
 func main() {
 	cfg.Init(
 		cfg.WithFlag(&cfg.Flag{Name: "a"}, &cfg.Flag{Name: "aa"}),
@@ -14,6 +15,6 @@ func main() {
 	log.Println(cfg.GetDuration("child[0].time"))
 	log.Println(cfg.GetSecond("child[1].time"))
 	log.Println(cfg.GetInt("a"))
-	log.Println(cfg.GetFloat64("aa"))
+	log.Println(cfg.GetInt("aa"))
 	log.Println(cfg.GetInt("b", 102))
 }
